mastermind_app/pip_requirements.txt
attrs==20.2.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
marshmallow==3.5.2
pytest-flask==1.0.0
